pyproject.toml
[tool.poetry]
name = "ralfs"
version = "1.0.0"
description = "Retrieval-Augmented Long-Form Summarization"
authors = ["Ashaduzzaman Sarker "]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ashaduzzaman-sarker/RALFS"
repository = "https://github.com/ashaduzzaman-sarker/RALFS"
documentation = "https://github.com/ashaduzzaman-sarker/RALFS/wiki"
keywords = ["nlp", "summarization", "retrieval", "rag", "long-form", "fid"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
packages = [{ include = "ralfs" }]  # ← Changed from "src/ralfs"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"

# Core ML
torch = "^2.4.0"
transformers = "4.44.2"
sentence-transformers = "3.0.1"
accelerate = "0.33.0"
peft = "0.11.1"
bitsandbytes = "0.43.3"  

# Data & Retrieval
datasets = "2.21.0"  
faiss-cpu = "1.8.0"
rank-bm25 = "0.2.2"
pyserini = "0.23.0"  

# ColBERT (use official Stanford NLP implementation)
# Install via: pip install git+https://github.com/stanford-futuredata/ColBERT.git
# colbert-ai = "0.2.20"  # ← Use this if available, else manual install

# NLP Tools
spacy = "3.7.5"
nltk = "3.8.1"
scikit-learn = "1.5.1"

# Evaluation
rouge-score = "0.1.2"
bert-score = "0.3.13"
evaluate = "0.4.2"

# Config & CLI
hydra-core = "1.3.2"
omegaconf = "2.3.0"
typer = {extras = ["all"], version = "0.12.3"}  
rich = "13.7.1"

# Experiment Tracking
wandb = "0.16.6"  

# Utilities
pandas = "2.2.2"
numpy = "1.26.4"
tqdm = "4.66.4"
pyyaml = "6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.0"
pytest-cov = "5.0.0"  
mypy = "1.11.0"
black = "24.8.0"
isort = "5.13.2"
ruff = "0.6.8"
pre-commit = "3.8.0"
jupyter = "1.0.0"
ipython = "8.26.0"
ipykernel = "6.29.5"

[tool.poetry.scripts]
ralfs = "ralfs.cli:app"  

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | __pycache__
  | build
  | dist
  | data
  | checkpoints
  | results
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.ruff]
target-version = "py310"
line-length = 100
select = ["E", "W", "F", "B", "C4", "UP", "I"]
ignore = ["E501", "B008"]  # E501: line too long, B008: function call in argument defaults
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "data",
    "checkpoints",
]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
warn_return_any = true
warn_redundant_casts = true
disallow_untyped_defs = false  # ← Set to true after adding type hints

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=ralfs --cov-report=term-missing --cov-report=html"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]